requirements.txt
matplotlib==3.5.3
numpy==1.21.5
opt_einsum==3.3.0
packaging==24.0
pandas==1.3.5
pynverse==0.1.4.6
pytorch_lightning==2.1.0
PyYAML==6.0.1
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.2
setuptools==65.6.3
tabulate==0.9.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
transformers==3.0.0
wget==3.2
